requirements.txt
numpy
pandas
scikit-learn
matplotlib
tensorflow
biopython
pysam
h5py